pyproject.toml
[project]
name = "downforacrossbot"
version = "0.1.0"
description = "Discord bot that sends downforacross crossword puzzle links"
requires-python = ">=3.14"
dependencies = [
    "discord-py",
    "python-dotenv",
    "httpx",
    "fastapi",
    "uvicorn[standard]",
    "pynacl",
    "pydantic-settings",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
    "ty>=0.0.1a25",
]


